+++ dwtools/TableOfReal_and_Permutation.cpp
@@ -31,10 +31,10 @@ autoTableOfReal TableOfReal_and_Permutation_permuteRows (TableOfReal me, Permuta
 		
 		autoTableOfReal him = TableOfReal_create (my numberOfRows, my numberOfColumns);
 
-		for (long i = 1; i <= thy numberOfElements; i ++) {
+		for (integer i = 1; i <= thy numberOfElements; i ++) {
 			TableOfReal_copyOneRowWithLabel (me, him.get(), thy p [i], i);
 		}
-		for (long j = 1; j <= my numberOfColumns; j ++) {
+		for (integer j = 1; j <= my numberOfColumns; j ++) {
 			TableOfReal_setColumnLabel (him.get(), j, my columnLabels [j]);
 		}
 		return him;
